@@ -1,4 +1,4 @@
-import { Component, OnInit, } from '@angular/core';
+import { booleanAttribute, Component, OnInit, } from '@angular/core';
 import { AuthService } from '../../services/auth.service';
 import { ParamService } from '../../services/param.service';
 import { Router } from '@angular/router';
@@ -10,10 +10,12 @@ import {
   FormsModule,
   ReactiveFormsModule,
   AbstractControl,
+  FormGroup,
 } from '@angular/forms';
 import {ErrorStateMatcher} from '@angular/material/core';
 import {MatInputModule} from '@angular/material/input';
 import {MatFormFieldModule} from '@angular/material/form-field';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-register',
@@ -34,9 +36,9 @@ export class RegisterComponent implements OnInit {
   };
 
   validateEmailMatch = (control: AbstractControl): {[key: string]: any} | null => {
-    const password = this.passTxt as string;
-    const passwordConfirm = control.value as string;
-    if (password !== passwordConfirm) {
+    const correo = this.correoTxt as string;
+    const correoConfirm = control.value as string;
+    if (correo !== correoConfirm) {
       return {emailMatch: true};
     }
     return null;
@@ -45,21 +47,42 @@ export class RegisterComponent implements OnInit {
   bgImgURL: string = "https://monsterslayersrepo.s3.amazonaws.com/Zones/CityBackground.gif";
   usernameTxt: string = "";
   correoTxt: string = "";
+  confirmCorreoTxt: string = "";
   passTxt: string = "";
   confirmPassTxt: string = "";
 
+  static regexStr ="(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])(?=.{8,})";
 
+  userFormControl = new FormControl('', [Validators.required]);
   emailFormControl = new FormControl('', [Validators.required, Validators.email]);
-  passFormControl = new FormControl('', [Validators.required, Validators.pattern('"^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$"')]);
-  confirmPassFormControl = new FormControl('', [Validators.required, this.validatePasswordMatch]);
   confirmEmailFormControl = new FormControl('', [Validators.required, this.validateEmailMatch]);
+  passFormControl = new FormControl('', [Validators.required, Validators.pattern(new RegExp(RegisterComponent.regexStr))]);
+  confirmPassFormControl = new FormControl('', [Validators.required, this.validatePasswordMatch]);
   matcher = new MyErrorStateMatcher();
+
+  registerform = new FormGroup([
+    this.userFormControl,
+    this.emailFormControl,
+    this.confirmEmailFormControl,
+    this.passFormControl,
+    this.confirmPassFormControl]);
+
+  emailErrotTxt = "";
+  emailErrotConfirmTxt = "";
+  passwordErrorTxt = "";
+  passwordErrorConfirmTxt = "";
+  passwordTypeTxt = true;
+  confirmPasswordTypeTxt = true;
+
+
 
   ngOnInit(): void {
   }
 
   createBtn(){
-    this.ExecuteRegister()
+    if(this.registerform.valid){
+      this.ExecuteRegister();
+    }
   }
 
   loginBtn(){
@@ -69,18 +92,34 @@ export class RegisterComponent implements OnInit {
   getBackground() {
     this.paramService.GetAllZones().subscribe(data => {
       var response = JSON.parse(data);
-      console.log(response);
     });
   }
 
   ExecuteRegister(){
-    console.log(this.confirmPassFormControl);
     this.authService.Register(this.usernameTxt, this.correoTxt, this.passTxt).subscribe(data => {
-      var response = JSON.parse(data);
-      console.log(response);
+      console.log(data);
+      Swal.fire({
+        title: "Usuario creado",
+        icon: "success",
+        showCancelButton: false,
+        confirmButtonColor: "#3f51b5",
+        confirmButtonText: "Aceptar"
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.router.navigate(['/login']);
+        }
+      });
+    },
+    error => {
+      if(error.status == 500){
+        Swal.fire({
+          title: error.error.message,
+          confirmButtonColor: "#3f51b5",
+          icon: "warning"
+        });
+      }
     });
   }
-
 }
 
 export class MyErrorStateMatcher implements ErrorStateMatcher {
